added events in dashboard and about us page

+++ public/js/dashboard.js
@@ -15,30 +15,55 @@ const showContent = (content) => {
     const deleteCourseBtn = document.getElementById('deleteCourseBtn');
 
     // getting id of course that needs to be deleted
-
-    deleteCourseBtn.addEventListener('click', (e) => {
-      const courseID = deleteCourseBtn.getAttribute('data-courseid');
-      axios({
-        method: 'DELETE',
-        url: '/api/v1/courses/' + courseID,
-      })
-        .then((res) => {
-          if (res.data.status === 'success') {
-            console.log(res.data);
-            alert('course deleted successfully');
-
-            location.assign('/dashboard');
-          }
+    if (deleteCourseBtn) {
+      deleteCourseBtn.addEventListener('click', (e) => {
+        const courseID = deleteCourseBtn.getAttribute('data-courseid');
+        axios({
+          method: 'DELETE',
+          url: '/api/v1/courses/' + courseID,
         })
-        .catch((err) => {
-          alert(err.response.data.message);
-        });
-    });
+          .then((res) => {
+            if (res.data.status === 'success') {
+              console.log(res.data);
+              alert('course deleted successfully');
+
+              location.assign('/dashboard');
+            }
+          })
+          .catch((err) => {
+            alert(err.response.data.message);
+          });
+      });
+    }
   } else if (content === 'event') {
     const template = eventTemplate();
     dashBoardContainer.innerHTML = '';
     let node = document.createRange().createContextualFragment(template);
     dashBoardContainer.appendChild(node);
+
+    const deleteEventBtn = document.getElementById('deleteEventBtn');
+
+    // getting id of course that needs to be deleted
+    if (deleteEventBtn) {
+      deleteEventBtn.addEventListener('click', (e) => {
+        const eventID = deleteEventBtn.getAttribute('data-eventID');
+        axios({
+          method: 'DELETE',
+          url: '/api/v1/events/' + eventID,
+        })
+          .then((res) => {
+            if (res.data.status === 'success') {
+              console.log(res.data);
+              alert('event deleted successfully');
+
+              location.assign('/dashboard');
+            }
+          })
+          .catch((err) => {
+            alert(err.response.data.message);
+          });
+      });
+    }
   }
 };
